+++ src/main/java/dev/jaxydog/lodestone/Lodestone.java
@@ -18,7 +18,6 @@ import dev.jaxydog.lodestone.api.Loaded;
 import dev.jaxydog.lodestone.impl.LoaderEnvironment;
 import dev.jaxydog.lodestone.impl.LoaderEnvironmentRegistry;
 import net.fabricmc.api.ModInitializer;
-import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.ApiStatus.Internal;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -47,7 +46,7 @@ public final class Lodestone implements ModInitializer {
      *
      * @since 1.0.0
      */
-    static final Logger LOGGER = LoggerFactory.getLogger(StringUtils.capitalize(MOD_ID));
+    static final Logger LOGGER = LoggerFactory.getLogger("Lodestone");
 
     /**
      * The mod's loader environment registry instance.